+++ backend/garpix_user/CHANGELOG.md
@@ -1,3 +1,8 @@
+### 3.8.0 (22.06.2023)
+
+- JWT token authorization added (see `Readme.md`)
+- Password restore by username fixed
+
 ### 3.7.1 (21.06.2023)
 
 - `USE_REGISTRATION` default value updated